Menu and Game improvements
Added an end condition to the game (with a menu that will pop up), needs to fix buttons here.

Added texture for buttons.

Added text fix for shop, so it will say "Selected Character" on the one currently selected.

Added a font to the game, and changed it to this everywhere